variables done
