+++ src/components/molecules/Variables/VariablesDescriptions/VariablesDescriptions.styled.tsx
