+++ src/components/molecules/Variables/VariablesDescriptions/index.ts
@@ -1,0 +1,2 @@
+export {default as TestDescription} from './TestDescription';
+export {default as TestSuiteDescription} from './TestSuiteDescription';
