+++ src/components/molecules/Variables/VariablesLists/TestsVariablesList/index.ts
@@ -1,0 +1,1 @@
+export {default} from './TestsVariablesList';
