+++ src/components/molecules/VariablesFormList/VariablesFormList.constants.tsx
@@ -1,0 +1,20 @@
+import {Variable} from '@models/variable';
+
+import {StyledPopoverContent} from './VariablesFormList.styled';
+
+export const popoverHelpContent = (
+  <StyledPopoverContent>If you wish to hide or obfuscate variables, select type Secret.</StyledPopoverContent>
+);
+
+export const emptyVariableObject: Variable = {
+  type: null,
+  key: '',
+  value: '',
+};
+
+export const typeOptions = [
+  {label: 'Basic', value: 0},
+  {label: 'Secret', value: 1},
+];
+
+export const duplicateKeyMessage = 'Duplicate key.';
