+++ src/utils/index.ts
@@ -7,4 +7,5 @@ export {
   checkApiEndpointProtocol,
   getQueryStringFromUrl,
   FinalizedApiEndpoint,
+  validateDuplicateValueByKey,
 } from './validate';
