+++ src/utils/validate.ts
@@ -1,14 +1,15 @@
-import { config } from '@constants/config';
-import { removeSpaceFromString, RemoveLastTrailingSlashFromString } from './strings';
+import {config} from '@constants/config';
+
+import {RemoveLastTrailingSlashFromString, removeSpaceFromString} from './strings';
 
 export const validateUrl = (url: string): boolean => {
   const pattern = new RegExp(
     '^(https?:\\/\\/)?' + // protocol
-    '((([a-z\\d]([a-z\\d-]*[a-z\\d])*)\\.?)+[a-z]{2,}|' + // domain name
-    '((\\d{1,3}\\.){3}\\d{1,3}))' + // OR ip (v4) address
-    '(\\:\\d+)?(\\/[-a-z\\d%_.~+]*)*' + // port and path
-    '(\\?[;&a-z\\d%_.~+=-]*)?' + // query string
-    '(\\#[-a-z\\d_]*)?$',
+      '((([a-z\\d]([a-z\\d-]*[a-z\\d])*)\\.?)+[a-z]{2,}|' + // domain name
+      '((\\d{1,3}\\.){3}\\d{1,3}))' + // OR ip (v4) address
+      '(\\:\\d+)?(\\/[-a-z\\d%_.~+]*)*' + // port and path
+      '(\\?[;&a-z\\d%_.~+=-]*)?' + // query string
+      '(\\#[-a-z\\d_]*)?$',
     'i'
   );
 
@@ -40,12 +41,11 @@ export const AppendApiVersionToEndpoint = (apiEndpoint: string) => {
 };
 
 export const FinalizedApiEndpoint = (apiEndpoint: string, storeData?: boolean) => {
- 
   const ApiEndpointWithoutSpace = removeSpaceFromString(apiEndpoint);
   const ApiEndpointWithoutTrailingSlash = RemoveLastTrailingSlashFromString(ApiEndpointWithoutSpace);
   const ApiEndpointWithProtocol = checkApiEndpointProtocol(ApiEndpointWithoutTrailingSlash);
   const ApiEndpointWithVersion = AppendApiVersionToEndpoint(ApiEndpointWithProtocol);
- 
+
   return ApiEndpointWithVersion;
 };
 
@@ -63,11 +63,25 @@ export const getApiEndpointOnPageLoad = () => {
     localStorage.removeItem(config.apiEndpoint);
     return;
   }
-   const finalUrl = FinalizedApiEndpoint(queryString, true);
- 
+  const finalUrl = FinalizedApiEndpoint(queryString, true);
+
   if (!finalUrl) {
     return;
   }
   localStorage.setItem(config.apiEndpoint, finalUrl);
+};
 
+export const validateDuplicateValueByKey = (value: string, list: any[], key: string) => {
+  let duplicateFlag = false;
+  // eslint-disable-next-line no-restricted-syntax
+  for (let i in list) {
+    if (value === list[i][key] && Boolean(value)) {
+      if (duplicateFlag) {
+        // eslint-disable-next-line prefer-promise-reject-errors
+        return true;
+      }
+      duplicateFlag = true;
+    }
+  }
+  return false;
 };
